Add form field id generation, remove duplicated ids
